Add files via upload
This project is a senario whereby Mr Lim is the new president of Singapore. 
Uncertainties arise as COVID-19 hits Singapore. He wants to know the severity of the impacts of the pandemic before deciding whether to tap into the past reserves and how effectively was the funds spent to deal with the pandemic

Get the other 2 large files >25MB from 
https://github.com/owid/covid-19-data/blob/master/public/data/cases_deaths/COVID-19%20Cases%20and%20deaths%20-%20WHO.csv
and 
https://github.com/owid/covid-19-data/blob/master/public/data/owid-covid-data.csv